refactored service engine to be a little easier to understand
