+++ packages/connectors/hull-outreach/test/integration/scenarios/outgoing-user-link-new-account-fail.test.js
@@ -41,9 +41,9 @@ test("send smart-notifier user update to outreach and link account", () => {
         ["info", "outgoing.job.start", expect.whatever(), {"jobName": "Outgoing Data", "type": "user"}],
         ["debug", "connector.service_api.call", expect.whatever(), {"method": "GET", "responseTime": expect.whatever(), "status": 200, "url": "/accounts/", "vars": {}}],
         ["debug", "connector.service_api.call",expect.whatever(), {"method": "POST", "responseTime": expect.whatever(), "status": 422, "url": "/accounts/", "vars": {}}],
-        ["error", "outgoing.account.skip", {"account_domain": "afterlife.com", "account_id": "5c0fd68ad884b4373800011a", "request_id": expect.whatever(), "subject_type": "account"}, { "data": expect.whatever(), operation: "post", "type": "Account", "error": "Outreach has rejected the objects being sent, please review attributes that you have in your filters to make sure that you've selected all the fields that outreach requires, if you think this is correct, please contact Hull support"}],
+        ["error", "outgoing.account.skip", {"account_domain": "afterlife.com", "account_id": "5c0fd68ad884b4373800011a", "request_id": expect.whatever(), "subject_type": "account"}, { "data": expect.whatever(), "type": "Account", "error": "Outreach has rejected the objects being sent, please review attributes that you have in your filters to make sure that you've selected all the fields that outreach requires, if you think this is correct, please contact Hull support"}],
         ["debug", "connector.service_api.call", expect.whatever(), {"method": "PATCH", "responseTime": expect.whatever(), "status": 200, "url": "/prospects/18", "vars": {}}],
-        ["info", "outgoing.user.success", {"request_id": expect.whatever(), "subject_type": "user", "user_email": "fettisbest@gmail.com", "user_id": "5bd329d5e2bcf3eeaf00009f"}, {"data": {"data": {"attributes": {"custom20": "in the afterlife"}, "id": 18, "type": "prospect"}}, "operation": "patch", "type": "Prospect"}],
+        ["info", "outgoing.user.success", {"request_id": expect.whatever(), "subject_type": "user", "user_email": "fettisbest@gmail.com", "user_id": "5bd329d5e2bcf3eeaf00009f"}, {"data": {"data": {"attributes": {"custom20": "in the afterlife"}, "id": 18, "type": "prospect"}}, "type": "Prospect"}],
         ["info", "incoming.user.success", {"request_id": expect.whatever(), "subject_type": "user", "user_email": "fettisbest@gmail.com", "user_anonymous_id": "outreach:18" }, { "data": expect.whatever(), type: "Prospect" }],
         ["info", "outgoing.job.success", expect.whatever(), {"jobName": "Outgoing Data", "type": "user"}]
       ],
